refactor: rename Utils.h to SvgUtils.h
