+++ tester/harmony/svg/src/main/cpp/utils/SvgUtils.h
